Problems: the add button doesn't actually add the hard coded data to state.  still working on resolving the issue.
installed redux logger for more detailed console info and redux-saga.
created a mutations file which has action types and the start of actions for tasks.
created a mock saga to start passing data from react to redux to initiate state changes.
updated the store for the index to have a combine reducer for all the data sets. It also has middleware and implementation - beggining of - sagas.
added in a map dispatch to the task list and a button to be able to add a new task.